+++ backend/worker/quivr_worker/files.py
@@ -1,11 +1,15 @@
 import hashlib
 import time
+from contextlib import contextmanager
 from pathlib import Path
+from tempfile import NamedTemporaryFile
 from typing import Any
 from uuid import UUID
 
 from quivr_api.logger import get_logger
 from quivr_core.files.file import FileExtension, QuivrFile
+
+from quivr_worker.utils import get_tmp_name
 
 logger = get_logger(__name__)
 
@@ -13,6 +17,43 @@ logger = get_logger(__name__)
 def compute_sha1(content: bytes):
     readable_hash = hashlib.sha1(content).hexdigest()
     return readable_hash
+
+
+@contextmanager
+def build_file(
+    file_data: bytes,
+    knowledge_id: UUID,
+    file_name: str,
+):
+    try:
+        # FIXME: @chloedia @AmineDiro
+        # We should decide if these checks should happen at API level or Worker level
+        # These checks should use Knowledge table (where we should store knowledge sha1)
+        # file_exists = file_already_exists()
+        # file_exists_in_brain = file_already_exists_in_brain(brain.brain_id)
+        # TODO(@aminediro) : Maybe use fsspec file to be agnostic to where files are stored :?
+        # We are reading the whole file to memory, which doesn't scale
+        tmp_name, base_file_name, file_extension = get_tmp_name(file_name)
+        tmp_file = NamedTemporaryFile(
+            suffix="_" + tmp_name,  # pyright: ignore reportPrivateUsage=none
+        )
+
+        file_sha1 = compute_sha1(file_data)
+        tmp_file.write(file_data)
+        tmp_file.flush()
+
+        file_instance = File(
+            knowledge_id=knowledge_id,
+            file_name=base_file_name,
+            tmp_file_path=Path(tmp_file.name),
+            file_size=len(file_data),
+            file_extension=file_extension,
+            file_sha1=file_sha1,
+        )
+        yield file_instance
+    finally:
+        # Code to release resource, e.g.:
+        tmp_file.close()
 
 
 class File:
