+++ backend/worker/tests/test_process_url_task.py
@@ -1,11 +1,11 @@
 import pytest
 
-from quivr_worker.crawl.crawler import CrawlWebsite
+from quivr_worker.parsers.crawler import URL, extract_from_url
 
 
 @pytest.mark.skip
 def test_crawl():
     url = "https://en.wikipedia.org/wiki/Python_(programming_language)"
-    crawl_website = CrawlWebsite(url=url)
-    extracted_content = crawl_website.process()
+    crawl_website = URL(url=url)
+    extracted_content = extract_from_url(crawl_website)
     assert len(extracted_content) > 1
